Initial commit
This creates a small bot that queries CloudWatch metrics for the number
of requests going to a particular load balancer.

Assign a Slack API token and load balancer name.

I've set it up as a Lambda function that is triggered once in the
morning, and once in the early evening.